rename component.tsx
