+++ webui/react/src/omnibar/Omnibar.module.scss
